requirements.txt
aiofiles==0.5.0
altair==4.1.0
aniso8601==7.0.0
argon2-cffi==21.1.0
astor==0.8.1
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.68.1
firebase-admin==5.0.2
gitdb==4.0.9
GitPython==3.1.24
google-api-core==1.31.2
google-api-python-client==2.21.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.40.0
h11==0.12.0
httplib2==0.19.1
idna==3.2
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.2
opencv-python==4.5.3.56
orjson==3.6.3
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
passlib==1.7.4
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.21
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.5
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
starlette==0.14.2
streamlit==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.1.0
uritemplate==3.0.1
urllib3==1.26.6
uvicorn==0.13.4
validators==0.18.2
watchdog==2.1.6
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.2
